refactor(visualization): move helpers to utils folder
