+++ src/ui/constants/visualizations/scatter.ts
@@ -7,8 +7,7 @@ import {
     PRIMITIVE_DATA_TYPES_AND_HIERARCHY,
 } from '../../constants/misc';
 import {prepareFieldToMeasureTransformation} from '../../units/wizard/utils/visualization';
-
-import {onMeasureAxisChange} from './line';
+import {onMeasureAxisChange} from '../../utils';
 
 export const SCATTER_VISUALIZATION: GraphShared['visualization'] = {
     id: WizardVisualizationId.Scatter,
